Version 1.0.2.7 (2021-01-06) * Bug fix: Coord.lon didn't expect a integer, it must be double.

+++ WeatherDemon/Models/OpenWeather.cs
@@ -4,7 +4,7 @@ namespace WeatherDemon.Models
 {
 	public class Coord
 	{
-		public int lon { get; set; }
+		public double lon { get; set; }
 		public double lat { get; set; }
 	}
